agent/server/requirements.txt
google-adk>=1.26.0
fastapi[all]>=0.135.1
uvicorn>=0.41.0
gunicorn>=25.1.0
python-dotenv>=1.2.2
google-cloud-firestore>=2.22.0
google-cloud-aiplatform>=1.86.0